refactor: set up golang project structure
