+++ src/internal/logger/zap.go
@@ -1,13 +1,13 @@
-package main
+package logger
 
 import (
 	"go.uber.org/zap"
 	//adapter "github.com/axiomhq/axiom-go/adapters/zap"
 )
 
-var logger *zap.SugaredLogger
+var Logger *zap.SugaredLogger
 
 func init() {
 	zapLogger, _ := zap.NewDevelopment() // zap.NewProduction()
-	logger = zapLogger.Sugar()
+	Logger = zapLogger.Sugar()
 }
